Update case.php
The refactored code continues with the array $modules that map the $op variable to the corresponding file name. Then, an if statement checks if the $op key is set in the $modules array. If it is, it includes the corresponding PHP file using include_once. If it's not, it exits the script with an error message.

This refactored code uses an array to map cases, using strict comparison and include_once instead of include.